+++ src/viam/examples/camera/example_camera.cpp
@@ -5,7 +5,7 @@
 
 #include <viam/sdk/components/camera.hpp>
 #include <viam/sdk/robot/client.hpp>
-#include <viam/sdk/robot/service.hpp>
+
 #include <viam/sdk/rpc/dial.hpp>
 
 int main() {
