+++ src/viam/examples/dial/example_dial.cpp
@@ -11,7 +11,7 @@
 
 #include <viam/sdk/components/generic.hpp>
 #include <viam/sdk/robot/client.hpp>
-#include <viam/sdk/robot/service.hpp>
+
 #include <viam/sdk/rpc/dial.hpp>
 
 using namespace viam::sdk;
