+++ src/viam/examples/modules/complex/client.cpp
@@ -13,7 +13,7 @@
 
 #include <viam/sdk/components/motor.hpp>
 #include <viam/sdk/robot/client.hpp>
-#include <viam/sdk/robot/service.hpp>
+
 #include <viam/sdk/rpc/dial.hpp>
 
 #include "gizmo/api.hpp"
