+++ src/viam/examples/motor/example_motor.cpp
@@ -6,7 +6,7 @@
 
 #include <viam/sdk/components/motor.hpp>
 #include <viam/sdk/robot/client.hpp>
-#include <viam/sdk/robot/service.hpp>
+
 #include <viam/sdk/rpc/dial.hpp>
 
 void print_motor_position(std::shared_ptr<viam::sdk::Motor> motor) {
