+++ src/viam/sdk/tests/mocks/mock_robot.cpp
@@ -8,6 +8,7 @@
 #include <robot/v1/robot.pb.h>
 
 #include <viam/sdk/common/proto_value.hpp>
+#include <viam/sdk/resource/stoppable.hpp>
 #include <viam/sdk/tests/test_utils.hpp>
 
 namespace viam {
@@ -16,10 +17,35 @@ namespace robot {
 using namespace viam::sdk;
 using common::v1::Pose;
 using common::v1::PoseInFrame;
+using common::v1::ResourceName;
 using viam::robot::v1::Discovery;
 using viam::robot::v1::DiscoveryQuery;
 using viam::robot::v1::FrameSystemConfig;
 using viam::robot::v1::Operation;
+
+namespace {
+std::vector<Name> registered_models_for_resource(const std::shared_ptr<Resource>& resource) {
+    std::string resource_type;
+    std::string resource_subtype;
+    std::vector<Name> resource_names;
+    for (const auto& kv : Registry::registered_models()) {
+        const std::shared_ptr<const ModelRegistration> reg = kv.second;
+        if (reg->api() == resource->api()) {
+            resource_type = reg->api().resource_type();
+            resource_subtype = reg->api().resource_subtype();
+        } else {
+            continue;
+        }
+
+        if (resource_subtype.empty()) {
+            resource_subtype = resource->name();
+        }
+
+        resource_names.push_back({{kRDK, resource_type, resource_subtype}, "", resource->name()});
+    }
+    return resource_names;
+}
+}  // namespace
 
 pose default_pose(int offset) {
     pose pose;
@@ -207,6 +233,102 @@ std::vector<FrameSystemConfig> mock_proto_config_response() {
     response.push_back(config);
     response.push_back(config1);
     return response;
+}
+
+MockRobotService::MockRobotService(const std::shared_ptr<ResourceManager>& manager, Server& server)
+    : ResourceServer(manager) {
+    server.register_service(this);
+    // register all managed resources with the appropriate resource servers.
+    for (const auto& resource : manager->resources()) {
+        server.add_resource(resource.second);
+    }
+}
+
+std::vector<ResourceName> MockRobotService::generate_metadata_() {
+    std::vector<ResourceName> metadata;
+    for (const auto& key_and_val : resource_manager()->resources()) {
+        for (const Name& name : registered_models_for_resource(key_and_val.second)) {
+            metadata.push_back(v2::to_proto(name));
+        }
+    }
+    return metadata;
+}
+
+::grpc::Status MockRobotService::ResourceNames(::grpc::ServerContext*,
+                                               const viam::robot::v1::ResourceNamesRequest* request,
+                                               viam::robot::v1::ResourceNamesResponse* response) {
+    if (!request) {
+        return ::grpc::Status(::grpc::StatusCode::INVALID_ARGUMENT,
+                              "Called [ResourceNames] without a request");
+    }
+
+    auto* p = response->mutable_resources();
+    for (const ResourceName& name : generate_metadata_()) {
+        *p->Add() = name;
+    }
+
+    return ::grpc::Status();
+}
+
+::grpc::Status MockRobotService::StopAll(::grpc::ServerContext*,
+                                         const ::viam::robot::v1::StopAllRequest* request,
+                                         ::viam::robot::v1::StopAllResponse*) {
+    const ResourceName r;
+    std::unordered_map<std::string, ProtoStruct> extra;
+    for (const auto& ex : request->extra()) {
+        const google::protobuf::Struct& struct_ = ex.params();
+        const ProtoStruct value_map = v2::from_proto(struct_);
+        const std::string name = ex.name().SerializeAsString();
+        extra.emplace(name, value_map);
+    }
+
+    grpc::StatusCode status = grpc::StatusCode::OK;
+    std::string status_message;
+
+    for (const auto& r : resource_manager()->resources()) {
+        const std::shared_ptr<Resource> resource = r.second;
+        const ResourceName rn = v2::to_proto(resource->get_resource_name());
+        const std::string rn_ = rn.SerializeAsString();
+        if (extra.find(rn_) != extra.end()) {
+            try {
+                Stoppable::stop_if_stoppable(resource, extra.at(rn_));
+            } catch (const std::runtime_error& err) {
+                try {
+                    status_message = err.what();
+                    Stoppable::stop_if_stoppable(resource);
+                } catch (std::runtime_error& err) {
+                    status_message = err.what();
+                    status = grpc::UNKNOWN;
+                } catch (...) {
+                    status_message = "unknown error";
+                    status = grpc::UNKNOWN;
+                }
+            }
+        } else {
+            try {
+                Stoppable::stop_if_stoppable(resource);
+            } catch (std::runtime_error& err) {
+                status_message = err.what();
+                status = grpc::UNKNOWN;
+            } catch (...) {
+                status_message = "unknown error";
+                status = grpc::UNKNOWN;
+            }
+        }
+    }
+
+    return grpc::Status(status, status_message);
+}
+
+std::shared_ptr<Resource> MockRobotService::resource_by_name(const Name& name) {
+    std::shared_ptr<Resource> r;
+    const std::lock_guard<std::mutex> lock(lock_);
+    auto resources = resource_manager()->resources();
+    if (resources.find(name.name()) != resources.end()) {
+        r = resources.at(name.name());
+    }
+
+    return r;
 }
 
 ::grpc::Status MockRobotService::FrameSystemConfig(
