+++ src/viam/sdk/tests/mocks/mock_robot.hpp
@@ -1,7 +1,13 @@
+#pragma once
+
+#include <common/v1/common.pb.h>
+#include <robot/v1/robot.grpc.pb.h>
 #include <robot/v1/robot.pb.h>
 
 #include <viam/sdk/common/pose.hpp>
-#include <viam/sdk/robot/service.hpp>
+#include <viam/sdk/resource/resource_server_base.hpp>
+#include <viam/sdk/robot/client.hpp>
+#include <viam/sdk/rpc/server.hpp>
 
 namespace viam {
 namespace sdktests {
@@ -17,10 +23,19 @@ using namespace viam::sdk;
 // service functionality. However, if/when these methods are updated in the actual RobotService_,
 // we should update the tests here accordingly.
 //
-class MockRobotService : public RobotService_ {
+class MockRobotService : public ResourceServer, public viam::robot::v1::RobotService::Service {
    public:
-    explicit MockRobotService(std::shared_ptr<ResourceManager> manager, Server& server)
-        : RobotService_(std::move(manager), server){};
+    MockRobotService(const std::shared_ptr<ResourceManager>& manager, Server& server);
+
+    std::shared_ptr<Resource> resource_by_name(const Name& name);
+
+    ::grpc::Status ResourceNames(::grpc::ServerContext* context,
+                                 const ::viam::robot::v1::ResourceNamesRequest* request,
+                                 ::viam::robot::v1::ResourceNamesResponse* response) override;
+
+    ::grpc::Status StopAll(::grpc::ServerContext* context,
+                           const ::viam::robot::v1::StopAllRequest* request,
+                           ::viam::robot::v1::StopAllResponse* response) override;
 
     ::grpc::Status FrameSystemConfig(
         ::grpc::ServerContext* context,
@@ -39,6 +54,10 @@ class MockRobotService : public RobotService_ {
     ::grpc::Status GetOperations(::grpc::ServerContext* context,
                                  const ::viam::robot::v1::GetOperationsRequest* request,
                                  ::viam::robot::v1::GetOperationsResponse* response) override;
+
+   private:
+    std::mutex lock_;
+    std::vector<common::v1::ResourceName> generate_metadata_();
 };
 
 pose default_pose(int offset = 0);
@@ -47,7 +66,7 @@ std::vector<viam::robot::v1::Operation> mock_proto_operations_response();
 std::vector<RobotClient::discovery> mock_discovery_response();
 std::vector<viam::robot::v1::Discovery> mock_proto_discovery_response();
 std::vector<Name> mock_resource_names_response();
-std::vector<ResourceName> mock_proto_resource_names_response();
+std::vector<common::v1::ResourceName> mock_proto_resource_names_response();
 std::vector<RobotClient::frame_system_config> mock_config_response();
 std::vector<viam::robot::v1::FrameSystemConfig> mock_proto_config_response();
 pose_in_frame mock_transform_response();
